refactor(projects): move rate limiters to feature folder
