+++ server/src/features/projects/middleware/rate-limiter.middleware.ts
@@ -1,0 +1,20 @@
+import rateLimit from "express-rate-limit";
+
+export const projectsRouteLimiter = rateLimit({
+  windowMs: 1 * 60 * 1000, // 1 minute
+  limit: 15, // 15 requests per window
+  standardHeaders: "draft-8", // idk wtf this is.
+  legacyHeaders: false,
+  message:
+    "Too many requests to projects route. Please try again after 1 minute.",
+  statusCode: 429,
+});
+
+export const projectTagsRouteLimiter = rateLimit({
+  windowMs: 1 * 60 * 1000, // 1 minute
+  limit: 5, // 5 requests per window
+  standardHeaders: "draft-8", // idk wtf this is.
+  legacyHeaders: false,
+  message: "Too many requests to tags route. Please try again after 1 minute.",
+  statusCode: 429,
+});
