+++ server/src/features/projects/projects.router.ts
@@ -1,7 +1,10 @@
 import express, { Request } from "express";
 import { API, AUTH } from "@data";
 import { validateRoles } from "@middleware";
-import { attachProjectCachePageService } from "./projects.middleware";
+import {
+  attachProjectCachePageService,
+  projectsRouteLimiter,
+} from "./middleware";
 import { IProjectFilter } from "./services";
 import { IProjectDetails } from "./types";
 
@@ -10,6 +13,7 @@ const { Guest, Student, Moderator, Administrator } = AUTH.ROLES_MAP;
 
 export const projectsRouter = express.Router();
 
+projectsRouter.use(projectsRouteLimiter);
 projectsRouter.use(attachProjectCachePageService);
 
 //  for testing
